requirements.txt
# ============================================================
# augment-lite-mcp 核心依賴
# ============================================================

# MCP Server
mcp>=1.1,<2

# HTTP 請求
requests>=2.32.0

# YAML 配置
pyyaml>=6.0

# 資料庫和檢索
duckdb>=1.4.0
rank-bm25>=0.2.2

# 向量檢索（需要先安裝 PyTorch）
# 注意：請使用 scripts/install_vector_deps.sh 安裝，或手動執行：
#   macOS: pip install torch torchvision torchaudio
#   Linux: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cpu
torch>=2.0.0
faiss-cpu>=1.9.0
sentence-transformers>=3.3.0
numpy>=1.26.0,<2.0.0  # 必須使用 1.x 版本，2.x 與 PyTorch/FAISS 不相容

# 環境變數管理
python-dotenv>=1.0.0

# ============================================================
# 開發工具（可選）
# ============================================================

# 代碼格式化和檢查
ruff>=0.14.0

# 測試
pytest>=8.4.0

# 類型檢查
mypy>=1.18.0

# 打包工具
build>=1.3.0